pyproject.toml
[project]
name = "nukki"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.1",
    "numpy>=2.4.1",
    "opencv-python>=4.13.0.90",
    "pillow>=12.1.0",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
]
